requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
discord.py==2.1.0
frozenlist==1.3.3
greenlet==2.0.2
hiredis==2.1.0
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
Pillow==9.4.0
pluggy==1.0.0
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==0.21.0
redis==4.5.5
SQLAlchemy==2.0.13
typing_extensions==4.4.0
yarl==1.9.2
